Changes to resolve a problem loading ICF files that use the new major.minor.subminor version scheme.
The problem is from ICF_ifstream::GetVersion function.  This function returns a long representation of the version string with the following code:
	version = (long)major *10000 + middle * 100 + minor;
So for example a version string of 1998.01.05 results in version = 19980105.  Later in the function there is a check and error for any version < 19980101.
Our new version string results in a version = 10100.  So it fails this check.

This change checks to see if the major is <1900 and if so returns the following version
	version =major*10000000000+middle * 100000 + minor;  

The return value of the function would was changed to DWORDLONG to prevent overflow in the large number cases. 
This also limits the major version number from 0-1900.  I made a note in IOVersion.h comments not to use major version beyond 1900.